+++ commands/common/http_push.py
@@ -89,11 +89,11 @@ class CommandHandler(CommandHandlerBase):
 
         if (self.link != None and self.link.response != None):
             Trace.log(TraceLevel.INFO, '')
-            Trace.log(TraceLevel.INFO, '[] HTTP Headers : {}'.format(self.link.response.getheaders()))
+            Trace.log(TraceLevel.INFO, '[] HTTP Headers : {}'.format(self.link.response.headers))
 
-        if (self.link != None and self.link.jsonData != None):
+        if (self.link != None and self.link.response != None):
             Trace.log(TraceLevel.INFO, '')
-            Trace.log(TraceLevel.INFO, '[] HTTP Data    : {}'.format(self.link.jsonData))
+            Trace.log(TraceLevel.INFO, '[] HTTP Data    : {}'.format(self.link.response.text))
 
         if (self.link != None and self.link.jsonData != None):
             Trace.log(TraceLevel.INFO, '')