+++ commands/systems/create_storagegroup.py
@@ -62,7 +62,7 @@ class CommandHandler(CommandHandlerBase):
         jsonRequest = CreateStorageGroupRequestProperties(redfishConfig, self.command, True)
 
         if jsonRequest is not None:
-            link = UrlAccess.process_request(redfishConfig, UrlStatus(url), 'POST', True, json.dumps(jsonRequest, indent=4))
+            link = UrlAccess.process_request(redfishConfig, UrlStatus(url), 'POST', True, jsonRequest)
 
             Trace.log(TraceLevel.INFO, '   -- {0: <14}: {1}'.format('Status', link.urlStatus))
             Trace.log(TraceLevel.INFO, '   -- {0: <14}: {1}'.format('Reason', link.urlReason))