+++ commands/systems/set_volume.py
@@ -83,7 +83,7 @@ class CommandHandler(CommandHandlerBase):
             volume = ''
 
         url = url + volume
-        link = UrlAccess.process_request(redfishConfig, UrlStatus(url), 'PATCH', True, json.dumps(JsonBuilder.getElement('main'), indent=4))
+        link = UrlAccess.process_request(redfishConfig, UrlStatus(url), 'PATCH', True, JsonBuilder.getElement('main'))
 
         Trace.log(TraceLevel.INFO, '   -- {0: <14}: {1}'.format('Status', link.urlStatus))
         Trace.log(TraceLevel.INFO, '   -- {0: <14}: {1}'.format('Reason', link.urlReason))